refactor(oauth, onboarding): CustomOAuth2User 계층 변경
