+++ backend/src/main/java/com/ssafy/dash/config/SecurityConfig.java
@@ -6,7 +6,7 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.web.SecurityFilterChain;
 
-import com.ssafy.dash.oauth.application.CustomOAuth2UserService;
+import com.ssafy.dash.oauth.presentation.security.CustomOAuth2UserService;
 
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
