+++ backend/src/main/java/com/ssafy/dash/oauth/presentation/security/CustomOAuth2UserService.java
@@ -1,4 +1,4 @@
-package com.ssafy.dash.oauth.application;
+package com.ssafy.dash.oauth.presentation.security;
 
 import java.util.Map;
 
@@ -13,8 +13,8 @@ import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.ssafy.dash.oauth.infrastructure.security.CustomOAuth2User;
 import com.ssafy.dash.oauth.application.dto.OAuthLoginResult;
+import com.ssafy.dash.oauth.application.OAuthTokenService;
 import com.ssafy.dash.user.application.UserService;
 
 @Service
