+++ backend/src/test/java/com/ssafy/dash/onboarding/presentation/OnboardingControllerTest.java
@@ -1,17 +1,14 @@
 package com.ssafy.dash.onboarding.presentation;
 
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.given;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
 import java.util.Collections;
 import java.util.Map;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.given;
+import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -26,18 +23,19 @@ import org.springframework.security.oauth2.client.authentication.OAuth2Authentic
 import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.test.web.servlet.MockMvc;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ssafy.dash.common.TestFixtures;
-import com.ssafy.dash.oauth.infrastructure.security.CustomOAuth2User;
-import com.ssafy.dash.oauth.domain.AuthFlowType;
 import com.ssafy.dash.oauth.application.dto.OAuthLoginResult;
+import com.ssafy.dash.oauth.domain.AuthFlowType;
+import com.ssafy.dash.oauth.presentation.security.CustomOAuth2User;
+import com.ssafy.dash.onboarding.application.OnboardingService;
 import com.ssafy.dash.onboarding.application.dto.RepositorySetupCommand;
 import com.ssafy.dash.onboarding.application.dto.RepositorySetupResult;
 import com.ssafy.dash.onboarding.presentation.dto.request.RepositorySetupRequest;
-import com.ssafy.dash.onboarding.application.OnboardingService;
-
-import org.mockito.Mockito;
 
 @WebMvcTest(OnboardingController.class)
 @AutoConfigureMockMvc(addFilters = false)
